requirements.txt
awscli==1.16.137
awsebcli==3.15.0
Django==2.2
django-bootstrap4==0.0.8
django-oidc-provider==0.7.0
django-phonenumber-field==2.3.1
phonenumbers==8.10.8
psycopg2==2.8.2
